+++ src/state/entity.ts
@@ -37,6 +37,7 @@ import './entity/ig_ENTITY_BounceSwitch'
 import './entity/ig_ENTITY_Destructible'
 import './entity/ig_ENTITY_EnemyCounter'
 import './entity/ig_ENTITY_DynamicPlatform'
+import './entity/ig_ENTITY_ItemDestruct'
 
 type EntityStateUnion = EntityStates[keyof EntityStates]
 
@@ -90,7 +91,7 @@ prestart(() => {
                     }
 
                     packet.states ??= {}
-                    packet.states[entity.netid] ??= {}
+                    packet.states[entity.netid] ??= {} as any
                     Object.assign(packet.states[entity.netid], state)
                 }
             }
@@ -130,7 +131,7 @@ export function createNetidStatic(
     z: number,
     settings: ig.Entity.Settings
 ): string {
-    let netid = typeId + encodeJsonSafeNumber(Number(`${x}${y}${z}`)) + settings.name
+    let netid = typeId + encodeJsonSafeNumber(Number(`${Math.floor(x)}${Math.floor(y)}${Math.floor(z)}`)) + settings.name
     if (settings.name) netid += settings.name
     return netid
 }